pyproject.toml
[project]
name = "multiagent-langgraph-circleci"
version = "0.1.0"
description = "A multi-agent research system using LangGraph for automated research and report generation"
readme = "README.md"
requires-python = ">=3.12,<3.13"                              
dependencies = [
    "langgraph>=0.0.10",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-aws>=0.0.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.3.19",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
# Set line length to 125
line-length = 125
target-version = "py312"

# Common directories to ignore
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist"
]

[tool.ruff.lint]
# Essential rules
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "W",    # pycodestyle warnings
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"

